move tests out of src to avoid bundling useless files
